Refactor SidePalette to PaletteEditor
